+++ lib/i18n_localizations_delegate.dart
@@ -7,19 +7,22 @@ class I18nLocalizationsDelegate
     extends LocalizationsDelegate<I18nLocalizations> {
   final List<Locale> supportedLocales;
   final String pathFile;
+  final bool haveCountryCode;
   static I18nLocalizations localizations;
 
-  I18nLocalizationsDelegate({this.supportedLocales, this.pathFile});
+  I18nLocalizationsDelegate(
+      {this.supportedLocales, this.pathFile, this.haveCountryCode = false});
 
   @override
   bool isSupported(Locale locale) {
-    return _languageCodeSuppotedLocales().contains(locale.languageCode);
+    return _languageCodeSupportedLocales().contains(_formatSupported(locale));
   }
 
   @override
   Future<I18nLocalizations> load(Locale locale) async {
     // I18nLocalizations class is where the JSON loading actually runs
-    localizations = I18nLocalizations(locale, supportedLocales, pathFile);
+    localizations =
+        I18nLocalizations(locale, supportedLocales, pathFile, haveCountryCode);
     await localizations.load();
     return localizations;
   }
@@ -31,8 +34,7 @@ class I18nLocalizationsDelegate
   localeResolutionCallback(Locale locale) {
     // Check if the current device locale is supported
     for (var supportedLocale in supportedLocales) {
-      if (supportedLocale.languageCode == locale.languageCode &&
-          supportedLocale.countryCode == locale.countryCode) {
+      if (_isSupportedLocales(supportedLocale, locale)) {
         return supportedLocale;
       }
     }
@@ -40,8 +42,20 @@ class I18nLocalizationsDelegate
     return supportedLocales.first;
   }
 
+  _isSupportedLocales(Locale supportedLocale, Locale locale) {
+    bool languageCode = supportedLocale.languageCode == locale.languageCode;
+    bool countryCode = supportedLocale.countryCode == locale.countryCode;
+    return haveCountryCode ? languageCode && countryCode : languageCode;
+  }
+
+  _formatSupported(Locale locale) {
+    return haveCountryCode
+        ? '${locale.languageCode}_${locale.countryCode}'
+        : locale.languageCode;
+  }
+
   // Map the languageCode from the supportedLocales
-  _languageCodeSuppotedLocales() {
-    return supportedLocales.map((locale) => locale.languageCode);
+  _languageCodeSupportedLocales() {
+    return supportedLocales.map((locale) => _formatSupported(locale));
   }
 }